+++ src/service/payment.ts
@@ -1,8 +1,0 @@
-/**
- * 決済サービス
- */
-import { service } from '@cinerino/domain';
-
-export import account = service.payment.account;
-export import creditCard = service.payment.creditCard;
-export import movietTicket = service.payment.movietTicket;