+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - repositoryを完全にCinerino化
+- serviceを完全にCinerino化
 
 ### Deprecated